server/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
pycryptodome
